requirements.txt
openai>=1.12.0
chromadb>=0.4.22
streamlit>=1.31.0
python-dotenv>=1.0.0
pypdf>=4.0.0
langchain-text-splitters>=0.0.1
